gpu/matrix_multiplication/pyproject.toml
[project]
name = "matrix-multiplication"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.2",
    "pyobjc-core>=11.1",
    "pyobjc-framework-cocoa>=11.1",
    "pyobjc-framework-metal>=11.1",
]
